move SceneControls functions to the SceneControls file
